move build options back into each "query", and fixed it
