+++ static/css/nl_interface.scss
@@ -86,9 +86,3 @@ h2,
 h3 {
   color: white;
 }
-
-#build-selector {
-  position: fixed;
-  top: 2em;
-  right: 2em;
-}
